Refactor MenuItem, add LegacyMenuItem and update usages
Refactored the MenuItem component to inherit from Button and simplified its implementation. Introduced LegacyMenuItem for previous dropdown and icon behaviors. Updated all usages in docs and stories to use new IconBefore/IconAfter/ChildContent patterns and migrated old MenuItem usages to LegacyMenuItem where appropriate. Added new Popover story, removed deprecated NewMenuItem, and updated navigation and data files to reflect these changes. Also added an 'Active' parameter to Button and improved label rendering.
