+++ BlueBlazor.Docs/Components/Layouts/MainLayout.razor
@@ -22,9 +22,9 @@
             </div>
         </nav>
 
-        <Button Variant="Shared.Variant.MenuItem" Label="Code on GitHub" Href="https://github.com/bruegmann/blue-blazor" Target="_blank">
+        <MenuItem Label="Code on GitHub" Href="https://github.com/bruegmann/blue-blazor" Target="_blank">
             <IconBefore><BiCodeSquare /></IconBefore>
-        </Button>
+        </MenuItem>
     </HeaderContent>
     <SideContent>
         <NavMenu />
