+++ BlueBlazor.Docs/Components/Layouts/NavMenu.razor
@@ -1,12 +1,9 @@
 ﻿<ul class="list-unstyled d-flex flex-column">
     <li>
-        <MenuItem Label="Home" Href="" IsActive="@IsActive("", true)">
-            <Icon>
+        <MenuItem Label="Home" Href="" Current="@IsActive("", true)">
+            <IconBefore>
                 <BiHouse />
-            </Icon>
-            <IconForActive>
-                <BiHouseFill />
-            </IconForActive>
+            </IconBefore>
         </MenuItem>
     </li>
 
@@ -20,16 +17,12 @@
                     @foreach (var item in group.Value.Where(i => i.Key == group.Key))
                     {
                         <li>
-                            <MenuItem Label="@item.Key" Href="@(item.Value + "/")" IsActive="@IsActive(item.Value + "/")">
-                                <Icon>
+                            <MenuItem Label="@item.Key" Href="@(item.Value + "/")" Current="@IsActive(item.Value + "/")">
+                                <IconBefore>
                                     @(componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.Icon ?? defaultIcon)
-                                </Icon>
-                                <IconForActive>
-                                    @(componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.IconForActive ??
-                                                            (componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.Icon ?? defaultIcon))
-                                </IconForActive>
+                                </IconBefore>
 
-                                <AfterLabelContent>
+                                <ChildContent>
                                     @if (componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.IsNew == true)
                                     {
                                         <span class="badge bg-primary ms-1 mt-1">New</span>
@@ -42,7 +35,7 @@
                                     {
                                         <span class="badge bg-warning ms-1 mt-1">Experimental</span>
                                     }
-                                </AfterLabelContent>
+                                </ChildContent>
                             </MenuItem>
                         </li>
                     }
@@ -50,16 +43,12 @@
                     @foreach (var item in group.Value.Where(i => i.Key != group.Key))
                     {
                         <li class="ms-3 border-start">
-                            <MenuItem Label="@item.Key" Href="@(item.Value + "/")" IsActive="@IsActive(item.Value + "/")">
-                                <Icon>
+                            <MenuItem Label="@item.Key" Href="@(item.Value + "/")" Current="@IsActive(item.Value + "/")">
+                                <IconBefore>
                                     @(componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.Icon ?? defaultIcon)
-                                </Icon>
-                                <IconForActive>
-                                    @(componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.IconForActive ??
-                                                            (componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.Icon ?? defaultIcon))
-                                </IconForActive>
+                                </IconBefore>
 
-                                <AfterLabelContent>
+                                <ChildContent>
                                     @if (componentIcons.Where(x => x.Key == item.Key).FirstOrDefault().Value?.IsNew == true)
                                     {
                                         <span class="badge bg-primary ms-1 mt-1">New</span>
@@ -72,7 +61,7 @@
                                     {
                                         <span class="badge bg-warning ms-1 mt-1">Experimental</span>
                                     }
-                                </AfterLabelContent>
+                                </ChildContent>
                             </MenuItem>
                         </li>
                     }
@@ -86,9 +75,9 @@
         <ul class="list-unstyled">
             <li>
                 <MenuItem Href="llms.txt" Label="llms.txt">
-                    <Icon>
+                    <IconBefore>
                         <BiBodyText />
-                    </Icon>
+                    </IconBefore>
                 </MenuItem>
             </li>
         </ul>
@@ -107,7 +96,9 @@
         {"Intro", new(@<BiPostcard />, @<BiPostcardFill />)},
         {"Layout", new(@<BiLayoutTextWindowReverse />)},
         {"Qr code gen", new(@<BiQrCode />)},
-        {"Menu item", new(@<BiApp />)},
+        {"Menu item", new(@<BiApp />, isNew: true)},
+        {"Menu item collapse", new(@<BiCardHeading />, isNew: true)},
+        {"Popover", new(@<BiMenuDown />, isNew: true)},
         {"Totp input", new(@<BiInputCursorText />)},
         {"Tab", new(@<BiSegmentedNav />)},
         {"Tabs", new(@<BiSegmentedNav />)},
