+++ BlueBlazor.Docs/Stories/Actions/Disposable.razor
@@ -3,23 +3,23 @@
     <div style="@($"width: {Width}px")">
         <Actions>
             <MenuItem Label="Child 1">
-                <Icon>🌍</Icon>
+                <IconBefore>🌍</IconBefore>
             </MenuItem>
 
             <MenuItem Label="Child 2">
-                <Icon>🌑</Icon>
+                <IconBefore>🌑</IconBefore>
             </MenuItem>
 
             <MenuItem Label="Child 3">
-                <Icon>☀️</Icon>
+                <IconBefore>☀️</IconBefore>
             </MenuItem>
 
             <MenuItem Label="Child 4">
-                <Icon>🚀</Icon>
+                <IconBefore>🚀</IconBefore>
             </MenuItem>
 
             <MenuItem Label="Child 5">
-                <Icon>🌌</Icon>
+                <IconBefore>🌌</IconBefore>
             </MenuItem>
         </Actions>
     </div>
