+++ BlueBlazor.Docs/Stories/Actions/InsideLayoutHeader.razor
@@ -5,23 +5,23 @@
         <Actions Class="bg-primary text-white">
             <Dropdown Label="Dropdown">
                 <MenuItem Label="Child 1">
-                    <Icon>🌍</Icon>
+                    <IconBefore>🌍</IconBefore>
                 </MenuItem>
 
                 <MenuItem Label="Child 2">
-                    <Icon>🌑</Icon>
+                    <IconBefore>🌑</IconBefore>
                 </MenuItem>
 
                 <MenuItem Label="Child 3">
-                    <Icon>☀️</Icon>
+                    <IconBefore>☀️</IconBefore>
                 </MenuItem>
 
                 <MenuItem Label="Child 4">
-                    <Icon>🚀</Icon>
+                    <IconBefore>🚀</IconBefore>
                 </MenuItem>
 
                 <MenuItem Label="Child 5">
-                    <Icon>🌌</Icon>
+                    <IconBefore>🌌</IconBefore>
                 </MenuItem>
             </Dropdown>
 
@@ -58,23 +58,23 @@
                 <Actions Class="bg-primary text-white">
                     <Dropdown Label="Dropdown">
                         <MenuItem Label="Child 1">
-                            <Icon>🌍</Icon>
+                            <IconBefore>🌍</IconBefore>
                         </MenuItem>
 
                         <MenuItem Label="Child 2">
-                            <Icon>🌑</Icon>
+                            <IconBefore>🌑</IconBefore>
                         </MenuItem>
 
                         <MenuItem Label="Child 3">
-                            <Icon>☀️</Icon>
+                            <IconBefore>☀️</IconBefore>
                         </MenuItem>
 
                         <MenuItem Label="Child 4">
-                            <Icon>🚀</Icon>
+                            <IconBefore>🚀</IconBefore>
                         </MenuItem>
 
                         <MenuItem Label="Child 5">
-                            <Icon>🌌</Icon>
+                            <IconBefore>🌌</IconBefore>
                         </MenuItem>
                     </Dropdown>
