+++ BlueBlazor.Docs/Stories/Button/WithIcon.razor
@@ -1,6 +1,6 @@
 ﻿@{
         /*docs {
-        "description": "Using the render fragments `IconBefore` or `IconAfter` will automatically set the [Bootstrap helper `icon-link`.](https://getbootstrap.com/docs/5.3/helpers/icon-link/#example)"
+        "description": "Using the render fragments `IconBefore` or `IconBeforeAfter` will automatically set the [Bootstrap helper `icon-link`.](https://getbootstrap.com/docs/5.3/helpers/icon-link/#example)"
     } */
 }
 
@@ -13,9 +13,9 @@
 </Button>
 
 <Button Label="Animate on hover" Class="icon-link-hover">
-    <IconAfter>
+    <IconBeforeAfter>
         <BiArrowRight />
-    </IconAfter>
+    </IconBeforeAfter>
 </Button>
 
 <Button Label="Custom hover transform" Class="icon-link-hover" style="--bs-icon-link-transform: translate3d(0, -.125rem, 0);">
@@ -25,7 +25,7 @@
 </Button>
 
 <Button Label="Emoji as icon" Class="icon-link-hover">
-    <IconAfter>
+    <IconBeforeAfter>
         <span>➡️</span>
-    </IconAfter>
+    </IconBeforeAfter>
 </Button>
