+++ BlueBlazor.Docs/Stories/Dropdown/BasicUsage.razor
@@ -1,8 +1,8 @@
 ﻿<Dropdown Label="Click to open">
     <MenuItem Label="First element">
-    <Icon>☀️</Icon>
+    <IconBefore>☀️</IconBefore>
     </MenuItem>
     <MenuItem Label="Second element">
-    <Icon>🚀</Icon>
+    <IconBefore>🚀</IconBefore>
     </MenuItem>
 </Dropdown>
