+++ BlueBlazor.Docs/Stories/Dropdown/Fallback.razor
@@ -1,8 +1,8 @@
 ﻿<Dropdown Label="Click to open" Fallback>
     <MenuItem Label="First element">
-    <Icon>☀️</Icon>
+    <IconBefore>☀️</IconBefore>
     </MenuItem>
     <MenuItem Label="Second element">
-    <Icon>🚀</Icon>
+    <IconBefore>🚀</IconBefore>
     </MenuItem>
 </Dropdown>
