+++ BlueBlazor.Docs/Stories/LegacyMenuItem/BasicUsage.razor
@@ -1,0 +1,41 @@
+﻿@{
+    /*docs {
+    "iframeUrl": "components/menu-item/stories/basic-usage"
+    } */
+}
+@layout EmptyLayout
+@page "/components/menu-item/stories/basic-usage"
+
+<Layout>
+    <SideContent>
+        <MenuItem Label="Hello">
+            <Icon>👋</Icon>
+        </MenuItem>
+        <MenuItem Label="Parent">
+            <Icon>☀️</Icon>
+            <DropdownContent>
+                <MenuItem Label="World">
+                    <Icon>🌍</Icon>
+                </MenuItem>
+                <MenuItem Label="Item 2">
+                    <Icon>🤔</Icon>
+                </MenuItem>
+                <MenuItem Label="Item 3">
+                    <Icon>☀️</Icon>
+                </MenuItem>
+                <MenuItem Label="Item 4">
+                    <Icon>🎉</Icon>
+                </MenuItem>
+                <MenuItem Label="Item 5">
+                    <Icon>😐</Icon>
+                </MenuItem>
+                <MenuItem Label="Item 6">
+                    <Icon>🤓</Icon>
+                </MenuItem>
+                <MenuItem Label="Item 7">
+                    <Icon>✅</Icon>
+                </MenuItem>
+            </DropdownContent>
+        </MenuItem>
+    </SideContent>
+</Layout>
