+++ BlueBlazor.Docs/Stories/MenuItem/BasicUsage.razor
@@ -1,41 +1,7 @@
-﻿@{
-    /*docs {
-    "iframeUrl": "components/menu-item/stories/basic-usage"
-    } */
-}
-@layout EmptyLayout
-@page "/components/menu-item/stories/basic-usage"
-
-<Layout>
-    <SideContent>
-        <MenuItem Label="Hello">
-            <Icon>👋</Icon>
-        </MenuItem>
-        <MenuItem Label="Parent">
-            <Icon>☀️</Icon>
-            <DropdownContent>
-                <MenuItem Label="World">
-                    <Icon>🌍</Icon>
-                </MenuItem>
-                <MenuItem Label="Item 2">
-                    <Icon>🤔</Icon>
-                </MenuItem>
-                <MenuItem Label="Item 3">
-                    <Icon>☀️</Icon>
-                </MenuItem>
-                <MenuItem Label="Item 4">
-                    <Icon>🎉</Icon>
-                </MenuItem>
-                <MenuItem Label="Item 5">
-                    <Icon>😐</Icon>
-                </MenuItem>
-                <MenuItem Label="Item 6">
-                    <Icon>🤓</Icon>
-                </MenuItem>
-                <MenuItem Label="Item 7">
-                    <Icon>✅</Icon>
-                </MenuItem>
-            </DropdownContent>
-        </MenuItem>
-    </SideContent>
-</Layout>
+﻿<MenuItem Label="Parent">
+    <IconBefore>
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16">
+            <path d="M12 8a4 4 0 1 1-8 0 4 4 0 0 1 8 0M8 0a.5.5 0 0 1 .5.5v2a.5.5 0 0 1-1 0v-2A.5.5 0 0 1 8 0m0 13a.5.5 0 0 1 .5.5v2a.5.5 0 0 1-1 0v-2A.5.5 0 0 1 8 13m8-5a.5.5 0 0 1-.5.5h-2a.5.5 0 0 1 0-1h2a.5.5 0 0 1 .5.5M3 8a.5.5 0 0 1-.5.5h-2a.5.5 0 0 1 0-1h2A.5.5 0 0 1 3 8m10.657-5.657a.5.5 0 0 1 0 .707l-1.414 1.415a.5.5 0 1 1-.707-.708l1.414-1.414a.5.5 0 0 1 .707 0m-9.193 9.193a.5.5 0 0 1 0 .707L3.05 13.657a.5.5 0 0 1-.707-.707l1.414-1.414a.5.5 0 0 1 .707 0m9.193 2.121a.5.5 0 0 1-.707 0l-1.414-1.414a.5.5 0 0 1 .707-.707l1.414 1.414a.5.5 0 0 1 0 .707M4.464 4.465a.5.5 0 0 1-.707 0L2.343 3.05a.5.5 0 1 1 .707-.707l1.414 1.414a.5.5 0 0 1 0 .708"></path>
+        </svg>
+    </IconBefore>
+</MenuItem>
