+++ BlueBlazor.Docs/Stories/MenuItemCollapse/BehaveAsCollapse.razor
@@ -1,7 +1,7 @@
 ﻿<MenuItemCollapse Label="Parent">
     <IconBefore><BiBackpack2Fill /></IconBefore>
     <ChildContent>
-        <NewMenuItem Label="Child 1" />
-        <NewMenuItem Label="Child 2" />
+        <MenuItem Label="Child 1" />
+        <MenuItem Label="Child 2" />
     </ChildContent>
 </MenuItemCollapse>
