+++ BlueBlazor.Docs/Stories/MenuItemCollapse/BehaveAsCollapseGroup.razor
@@ -1,9 +1,9 @@
 ﻿<div class="blue-collapse-group">
-    <NewMenuItem Label="Parent">
+    <MenuItem Label="Parent">
         <IconBefore>
             <BiDropletFill />
         </IconBefore>
-    </NewMenuItem>
+    </MenuItem>
 
     <MenuItemCollapse>
         <NewMenuItem Label="Child 1" />
