--- BlueBlazor.Docs/Stories/NewMenuItem/BehaveAsPopoverGroup.razor
@@ -1,12 +1,0 @@
-﻿<CreateId>
-    <NewMenuItem Label="Open popover" PopoverTarget="context">
-        <IconAfter>
-            <BiChevronDown />
-        </IconAfter>
-    </NewMenuItem>
-
-    <MenuItemPopover Id="context">
-        <NewMenuItem Label="Sub item 1" />
-        <NewMenuItem Label="Sub longer item 2" />
-    </MenuItemPopover>
-</CreateId>
