+++ BlueBlazor.Docs/Stories/Popover/BasicUsage.razor
@@ -1,0 +1,12 @@
+﻿<CreateId>
+    <MenuItem Label="Open popover" PopoverTarget="context">
+        <IconAfter>
+            <BiChevronDown />
+        </IconAfter>
+    </MenuItem>
+
+    <Popover Id="context">
+        <MenuItem Label="Sub item 1" />
+        <MenuItem Label="Sub longer item 2" />
+    </Popover>
+</CreateId>
