+++ BlueBlazor.Docs/Stories/TextInput/SearchInputGroup.razor
@@ -1,5 +1,5 @@
 ﻿<div class="blue-input-group input-group">
     <TextInput Label="Search" LabelHidden LabelClass="input-group-text">
-        <AfterLabelContent><BiSearch aria-hidden="true" /></AfterLabelContent>
+        <ButtonContent><BiSearch aria-hidden="true" /></ButtonContent>
     </TextInput>
 </div>
