+++ BlueBlazor.Docs/wwwroot/data/pages-data.json
@@ -1,1 +1,1 @@
-{"Action menu":{"Action menu":"components/action-menu"},"Actions":{"Actions":"components/actions"},"Alert section":{"Alert section":"components/alert-section"},"Button":{"Button":"components/button","Button or a":"components/button-or-a"},"Chevron":{"Chevron":"components/chevron"},"Collapse":{"Collapse":"components/collapse"},"Create id":{"Create id":"components/create-id"},"Dialog provider":{"Dialog opener":"components/dialog-opener","Dialog provider":"components/dialog-provider","Modal dialog":"components/modal-dialog","Offcanvas dialog":"components/offcanvas-dialog"},"Dropdown":{"Dropdown":"components/dropdown"},"Field groups":{"Field group":"components/field-group","Field groups":"components/field-groups"},"Header title":{"Header title":"components/header-title"},"Heading":{"Heading":"components/heading"},"Inline edit":{"Inline edit":"components/inline-edit"},"Intro":{"Intro":"components/intro"},"Layout":{"Layout":"components/layout"},"Menu item":{"Menu item":"components/menu-item","Menu item collapse":"components/menu-item-collapse","Menu item popover":"components/menu-item-popover","New menu item":"components/new-menu-item"},"Monaco editor":{"Monaco editor":"components/monaco-editor"},"Odometer":{"Odometer":"components/odometer"},"Outside":{"Outside":"components/outside"},"Page script":{"Page script":"components/page-script"},"Qr code gen":{"Qr code gen":"components/qr-code-gen"},"Sidebar menu":{"Sidebar menu":"components/sidebar-menu"},"Simple layout":{"Simple layout":"components/simple-layout"},"Tabs":{"Tab":"components/tab","Tabs":"components/tabs"},"Text input":{"Text input":"components/text-input"},"Tooltip":{"Tooltip":"components/tooltip"},"Totp input":{"Totp input":"components/totp-input"},"Tui editor":{"Tui editor":"components/tui-editor"}}
+{"Action menu":{"Action menu":"components/action-menu"},"Actions":{"Actions":"components/actions"},"Alert section":{"Alert section":"components/alert-section"},"Button":{"Button":"components/button","Button or a":"components/button-or-a"},"Chevron":{"Chevron":"components/chevron"},"Collapse":{"Collapse":"components/collapse"},"Create id":{"Create id":"components/create-id"},"Dialog provider":{"Dialog opener":"components/dialog-opener","Dialog provider":"components/dialog-provider","Modal dialog":"components/modal-dialog","Offcanvas dialog":"components/offcanvas-dialog"},"Dropdown":{"Dropdown":"components/dropdown"},"Field groups":{"Field group":"components/field-group","Field groups":"components/field-groups"},"Header title":{"Header title":"components/header-title"},"Heading":{"Heading":"components/heading"},"Inline edit":{"Inline edit":"components/inline-edit"},"Intro":{"Intro":"components/intro"},"Layout":{"Layout":"components/layout"},"Menu item":{"Legacy menu item":"components/legacy-menu-item","Menu item":"components/menu-item","Menu item collapse":"components/menu-item-collapse"},"Monaco editor":{"Monaco editor":"components/monaco-editor"},"Odometer":{"Odometer":"components/odometer"},"Outside":{"Outside":"components/outside"},"Page script":{"Page script":"components/page-script"},"Popover":{"Popover":"components/popover"},"Qr code gen":{"Qr code gen":"components/qr-code-gen"},"Sidebar menu":{"Sidebar menu":"components/sidebar-menu"},"Simple layout":{"Simple layout":"components/simple-layout"},"Tabs":{"Tab":"components/tab","Tabs":"components/tabs"},"Text input":{"Text input":"components/text-input"},"Tooltip":{"Tooltip":"components/tooltip"},"Totp input":{"Totp input":"components/totp-input"},"Tui editor":{"Tui editor":"components/tui-editor"}}
