--- BlueBlazor/Components/MenuItem/NewMenuItem.razor
@@ -1,4 +1,0 @@
-﻿@namespace BlueBlazor.Components
-@inherits Button
-
-<Button @ref="ButtonRef" />
